+++ .gitignore
@@ -2,6 +2,9 @@
 .DS_Store
 *.swp
 *.swo
+*.swl
+*.swm
+*.swn
 .vscode
 .idea
