+++ server/start.go
@@ -57,12 +57,18 @@ func StartCmd(ctx *Context, appCreator AppCreator) *cobra.Command {
 func startStandAlone(ctx *Context, appCreator AppCreator) error {
 	addr := viper.GetString(flagAddress)
 	home := viper.GetString("home")
-	traceStore := viper.GetString(flagTraceStore)
+	traceWriterFile := viper.GetString(flagTraceStore)
 
-	app, err := appCreator(home, ctx.Logger, traceStore)
+	db, err := openDB(home)
 	if err != nil {
 		return err
 	}
+	traceWriter, err := openTraceWriter(traceWriterFile)
+	if err != nil {
+		return err
+	}
+
+	app := appCreator(ctx.Logger, db, traceWriter)
 
 	svr, err := server.NewServer(addr, "socket", app)
 	if err != nil {
@@ -91,12 +97,18 @@ func startStandAlone(ctx *Context, appCreator AppCreator) error {
 func startInProcess(ctx *Context, appCreator AppCreator) (*node.Node, error) {
 	cfg := ctx.Config
 	home := cfg.RootDir
-	traceStore := viper.GetString(flagTraceStore)
+	traceWriterFile := viper.GetString(flagTraceStore)
 
-	app, err := appCreator(home, ctx.Logger, traceStore)
+	db, err := openDB(home)
 	if err != nil {
 		return nil, err
 	}
+	traceWriter, err := openTraceWriter(traceWriterFile)
+	if err != nil {
+		return nil, err
+	}
+
+	app := appCreator(ctx.Logger, db, traceWriter)
 
 	nodeKey, err := p2p.LoadOrGenNodeKey(cfg.NodeKeyFile())
 	if err != nil {
