+++ server/test_helpers.go
@@ -36,9 +36,9 @@ func FreeTCPAddr() (addr, port string, err error) {
 	return
 }
 
-// setupViper creates a homedir to run inside,
+// SetupViper creates a homedir to run inside,
 // and returns a cleanup function to defer
-func setupViper(t *testing.T) func() {
+func SetupViper(t *testing.T) func() {
 	rootDir, err := ioutil.TempDir("", "mock-sdk-cmd")
 	require.Nil(t, err)
 	viper.Set(cli.HomeFlag, rootDir)
